[refactor] AdmTradeController > AdminController에 병합
